add 'issues' page

+++ src/main/java/com/strelnikov/bugtracker/dao/ProjectRepository.java
@@ -3,5 +3,5 @@ package com.strelnikov.bugtracker.dao;
 import com.strelnikov.bugtracker.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository  extends JpaRepository<Project,Integer> {
+public interface ProjectRepository  extends JpaRepository<Project,Long> {
 }
